ITEM + BUTTON
The buttons have relevant methods
-OnHover
-OnSelect
-ButtonMethod
-SetText
There's a lot of other stuff going on with that script (specifically within the update function) but if there's better existing ways of navigating and having these activate, then I can go back and delete all of that.

The Item has relevant methods
-SetText
-SetImage
-SetScale

The SetScale is just there so we can make items of different sizes. This one is made to look like a polaroid because I assumed that's what we wanted, but making any that don't have outlines/look any different is 0 work.